+++ git_machete/cmd.py
@@ -135,6 +135,8 @@ class CommandLineContext:
         self.opt_no_detect_squash_merges: bool = False
         self.opt_no_edit_merge: bool = False
         self.opt_no_interactive_rebase: bool = False
+        self.opt_no_push: Optional[bool] = None
+        self.opt_no_push_untracked: Optional[bool] = None
         self.opt_onto: Optional[str] = None
         self.opt_override_to: Optional[str] = None
         self.opt_override_to_inferred: bool = False
@@ -149,7 +151,16 @@ class CommandLineContext:
         self.opt_yes: bool = False
 
 
-def ask_if(cli_ctxt: CommandLineContext, msg: str, opt_yes_msg: Optional[str], apply_fmt: bool = True) -> str:
+def ask_if(
+    cli_ctxt: CommandLineContext,
+    msg: str,
+    opt_yes_msg: Optional[str],
+    apply_fmt: bool = True,
+    override: Optional[str] = None,
+) -> str:
+    if override:
+        print((fmt(msg) if apply_fmt else msg) + f" {override} [override]")
+        return override
     if cli_ctxt.opt_yes and opt_yes_msg:
         print(fmt(opt_yes_msg) if apply_fmt else opt_yes_msg)
         return 'y'
@@ -2254,7 +2265,8 @@ def handle_untracked_branch(cli_ctxt: CommandLineContext, new_remote: str, b: st
     if not commit_sha_by_revision(cli_ctxt, rb, prefix="refs/remotes/"):
         msg = f"Push untracked branch {bold(b)} to {bold(new_remote)}?" + pretty_choices('y', 'N', 'q', 'yq', other_remote_choice)
         opt_yes_msg = f"Pushing untracked branch {bold(b)} to {bold(new_remote)}..."
-        ans = ask_if(cli_ctxt, msg, opt_yes_msg)
+        override = "N" if cli_ctxt.opt_no_push or cli_ctxt.opt_no_push_untracked else None
+        ans = ask_if(cli_ctxt, msg, opt_yes_msg, override=override)
         if ans in ('y', 'yes', 'yq'):
             push(cli_ctxt, new_remote, b)
             if ans == 'yq':
@@ -2265,6 +2277,8 @@ def handle_untracked_branch(cli_ctxt: CommandLineContext, new_remote: str, b: st
         elif ans in ('q', 'quit'):
             raise StopTraversal
         return
+
+    relation = get_relation_to_remote_counterpart(cli_ctxt, b, rb)
 
     message = {
         IN_SYNC_WITH_REMOTE:
@@ -2277,9 +2291,9 @@ def handle_untracked_branch(cli_ctxt: CommandLineContext, new_remote: str, b: st
             f"Branch {bold(b)} is untracked, it diverged from its remote counterpart candidate {bold(rb)}, and has {bold('older')} commits than {bold(rb)}.",
         DIVERGED_FROM_AND_NEWER_THAN_REMOTE:
             f"Branch {bold(b)} is untracked, it diverged from its remote counterpart candidate {bold(rb)}, and has {bold('newer')} commits than {bold(rb)}."
-    }
+    }[relation]
 
-    prompt = {
+    msg, opt_yes_msg = {
         IN_SYNC_WITH_REMOTE: (
             f"Set the remote of {bold(b)} to {bold(new_remote)} without pushing or pulling?" + pretty_choices('y', 'N', 'q', 'yq', other_remote_choice),
             f"Setting the remote of {bold(b)} to {bold(new_remote)}..."
@@ -2300,22 +2314,28 @@ def handle_untracked_branch(cli_ctxt: CommandLineContext, new_remote: str, b: st
             f"Push branch {bold(b)} with force-with-lease to {bold(new_remote)}?" + pretty_choices('y', 'N', 'q', 'yq', other_remote_choice),
             f"Pushing branch {bold(b)} with force-with-lease to {bold(new_remote)}..."
         )
-    }
+    }[relation]
 
-    yes_actions: Dict[int, Callable[[], None]] = {
+    override = {
+        IN_SYNC_WITH_REMOTE: None,
+        BEHIND_REMOTE: None,
+        AHEAD_OF_REMOTE: "N" if cli_ctxt.opt_no_push or cli_ctxt.opt_no_push_untracked else None,
+        DIVERGED_FROM_AND_OLDER_THAN_REMOTE: None,
+        DIVERGED_FROM_AND_NEWER_THAN_REMOTE: "N" if cli_ctxt.opt_no_push or cli_ctxt.opt_no_push_untracked else None,
+    }[relation]
+
+    yes_action: Callable[[], None] = {
         IN_SYNC_WITH_REMOTE: lambda: set_upstream_to(cli_ctxt, rb),
         BEHIND_REMOTE: lambda: pull_ff_only(cli_ctxt, new_remote, rb),
         AHEAD_OF_REMOTE: lambda: push(cli_ctxt, new_remote, b),
         DIVERGED_FROM_AND_OLDER_THAN_REMOTE: lambda: reset_keep(cli_ctxt, rb),
         DIVERGED_FROM_AND_NEWER_THAN_REMOTE: lambda: push(cli_ctxt, new_remote, b, force_with_lease=True)
-    }
+    }[relation]
 
-    relation = get_relation_to_remote_counterpart(cli_ctxt, b, rb)
-    print(message[relation])
-    msg, opt_yes_msg = prompt[relation]
-    ans = ask_if(cli_ctxt, msg, opt_yes_msg)
+    print(message)
+    ans = ask_if(cli_ctxt, msg, opt_yes_msg, override=override)
     if ans in ('y', 'yes', 'yq'):
-        yes_actions[relation]()
+        yes_action()
         if ans == 'yq':
             raise StopTraversal
         flush_caches()
@@ -2500,7 +2520,8 @@ def traverse(cli_ctxt: CommandLineContext) -> None:
                 ans = ask_if(
                     cli_ctxt,
                     f"Push {bold(b)} to {bold(remote)}?" + pretty_choices('y', 'N', 'q', 'yq'),
-                    f"Pushing {bold(b)} to {bold(remote)}..."
+                    f"Pushing {bold(b)} to {bold(remote)}...",
+                    override="N" if cli_ctxt.opt_no_push else None,
                 )
                 if ans in ('y', 'yes', 'yq'):
                     push(cli_ctxt, remote, b)
@@ -2536,7 +2557,8 @@ def traverse(cli_ctxt: CommandLineContext) -> None:
                     f"Branch {bold(b)} diverged from (and has newer commits than) its remote counterpart {bold(rb)}.\n"
                     f"Push {bold(b)} with force-with-lease to {bold(remote)}?" + pretty_choices('y', 'N', 'q', 'yq'),
                     f"Branch {bold(b)} diverged from (and has newer commits than) its remote counterpart {bold(rb)}.\n"
-                    f"Pushing {bold(b)} with force-with-lease to {bold(remote)}..."
+                    f"Pushing {bold(b)} with force-with-lease to {bold(remote)}...",
+                    override="N" if cli_ctxt.opt_no_push else None,
                 )
                 if ans in ('y', 'yes', 'yq'):
                     push(cli_ctxt, remote, b, force_with_lease=True)
@@ -3313,6 +3335,7 @@ long_docs: Dict[str, str] = {
     "traverse": """
         <b>Usage: git machete traverse [-F|--fetch] [-l|--list-commits] [-M|--merge]
                                        [-n|--no-edit-merge|--no-interactive-rebase] [--no-detect-squash-merges]
+                                       [--[no-]push] [--[no-]push-untracked]
                                        [--return-to=WHERE] [--start-from=WHERE] [-w|--whole] [-W] [-y|--yes]</b>
 
         Traverses the branch dependency in pre-order (i.e. simply in the order as they occur in the definition file) and for each branch:
@@ -3352,6 +3375,10 @@ long_docs: Dict[str, str] = {
 
           <b>--no-interactive-rebase</b>      If updating by rebase, run `git rebase` in non-interactive mode (without `-i/--interactive` flag).
                                        Not allowed if updating by merge.
+
+          <b>--no-push</b>                  Do not push branches to remote, reenable via --push.
+
+          <b>--no-push-untracked</b>        Do not push branches without existing remote tracking branch to remote, reenable via --push-untracked.
 
           <b>--return-to=WHERE</b>            Specifies the branch to return after traversal is successfully completed; WHERE can be `here` (the current branch at the moment when traversal starts),
                                        `nearest-remaining` (nearest remaining branch in case the `here` branch has been slid out by the traversal)
@@ -3496,6 +3523,14 @@ def launch(orig_args: List[str]) -> None:
                 cli_ctxt.opt_no_edit_merge = True
             elif opt == "--no-interactive-rebase":
                 cli_ctxt.opt_no_interactive_rebase = True
+            elif opt == "--no-push":
+                cli_ctxt.opt_no_push = True
+            elif opt == "--push":
+                cli_ctxt.opt_no_push = False
+            elif opt == "--no-push-untracked":
+                cli_ctxt.opt_no_push_untracked = True
+            elif opt == "--push-untracked":
+                cli_ctxt.opt_no_push_untracked = False
             elif opt in ("-o", "--onto"):
                 cli_ctxt.opt_onto = arg
             elif opt == "--override-to":
@@ -3773,6 +3808,7 @@ def launch(orig_args: List[str]) -> None:
         elif cmd == "traverse":
             traverse_long_opts = ["fetch", "list-commits", "merge",
                                   "no-detect-squash-merges", "no-edit-merge", "no-interactive-rebase",
+                                  "no-push", "push", "no-push-untracked", "push-untracked",
                                   "return-to=", "start-from=", "whole", "yes"]
             expect_no_param(parse_options(args, "FlMnWwy", traverse_long_opts))
             if cli_ctxt.opt_start_from not in ("here", "root", "first-root"):